long

+++ src/main/java/Main.java
@@ -3,13 +3,13 @@ import ru.netology.stats.StatsService;
 public class Main {
     public static void main(String[] args) {
         StatsService service = new StatsService();
-        int[] sum = {
+        long [] sum = {
                 8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18
         };
-        int sumAllMonth = service.getSum(sum);
+        long sumAllMonth = service.getSum(sum);
         System.out.println(sumAllMonth);
 
-        int midSum = service.midSum(sum);
+        long midSum = service.midSum(sum);
         System.out.println(midSum);
 
         int pickSales = service.pickSales(sum);
