+++ src/main/java/ru/netology/stats/StatsService.java
@@ -1,20 +1,20 @@
 package ru.netology.stats;
 
 public class StatsService {
-    public int getSum(int[] sum) {
-        int sumAllMonth = 0;
+    public long getSum(long[] sum) {
+        long sumAllMonth = 0;
         for (int i = 0; i < sum.length; i++) {
             sumAllMonth = sumAllMonth + sum[i];
         }
         return sumAllMonth;
     }
 
-    public int midSum(int[] sum) {
-        int getSum = this.getSum(sum);
+    public long midSum(long[] sum) {
+        long getSum = this.getSum(sum);
         return (int) getSum / sum.length;
     }
 
-    public int pickSales(int[] sum) {
+    public int pickSales(long[] sum) {
         int pickMonth = 0;
         for (int i = 0; i < sum.length; i++) {
             if (sum[i] >= sum[pickMonth]) {
@@ -24,7 +24,7 @@ public class StatsService {
         return pickMonth;
     }
 
-    public int minSales(int[] sum) {
+    public int minSales(long[] sum) {
         int minMonth = 0;
         for (int i = 0; i < sum.length; i++) {
             if (sum[i] <= sum[minMonth]) {
@@ -34,10 +34,10 @@ public class StatsService {
         return minMonth;
     }
 
-    public int countBelowAverageMonth(int[] sum) {
-        int avg = this.midSum(sum);
+    public int countBelowAverageMonth(long[] sum) {
+        long avg = this.midSum(sum);
         int belowAvgCount = 0;
-        for (int m : sum) {
+        for (long m : sum) {
             if (m < avg) {
                 belowAvgCount++;
             }
@@ -46,10 +46,10 @@ public class StatsService {
 
     }
 
-    public int countAboveAverageMonth(int[] sum) {
-        int avg = this.midSum(sum);
+    public int countAboveAverageMonth(long[] sum) {
+        long avg = this.midSum(sum);
         int aboveAvgCount = 0;
-        for (int m : sum) {
+        for (long m : sum) {
             if (m > avg) {
                 aboveAvgCount++;
             }
